Small Bug Fix and Streamline Fading In of divs
Fixed bug where an icon was square when it was supposed to be circular.
Added 2 new js files to keep "div fades in when scrolled down enough" in it's own place. Now can call on a function for that kind of fading instead of copy/pasting a chunk of code each time.